Split inline
