+++ frictionless/plugins/inline/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import InlinePlugin
+from .dialect import InlineDialect
+from .parser import InlineParser
