+++ frictionless/plugins/inline/parser.py
@@ -1,108 +1,6 @@
-import typing
-from ..exception import FrictionlessException
-from ..metadata import Metadata
-from ..dialect import Dialect
-from ..plugin import Plugin
-from ..parser import Parser
-from .. import errors
-
-
-# NOTE:
-# Consider renaming keys/data_keys to labels due to dict.keys conflict
-
-
-# Plugin
-
-
-class InlinePlugin(Plugin):
-    """Plugin for Inline
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless.plugins.inline import InlinePlugin`
-
-    """
-
-    code = "inline"
-
-    def create_file(self, file):
-        if not file.scheme and not file.format and file.memory:
-            if not hasattr(file.data, "read"):
-                types = (list, typing.Iterator, typing.Generator)
-                if callable(file.data) or isinstance(file.data, types):
-                    file.scheme = ""
-                    file.format = "inline"
-                    return file
-
-    def create_dialect(self, resource, *, descriptor):
-        if resource.format == "inline":
-            return InlineDialect(descriptor)
-
-    def create_parser(self, resource):
-        if resource.format == "inline":
-            return InlineParser(resource)
-
-
-# Dialect
-
-
-class InlineDialect(Dialect):
-    """Inline dialect representation
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless.plugins.inline import InlineDialect`
-
-    Parameters:
-        descriptor? (str|dict): descriptor
-        keys? (str[]): a list of strings to use as data keys
-        keyed? (bool): whether data rows are keyed
-
-    Raises:
-        FrictionlessException: raise any error that occurs during the process
-
-    """
-
-    def __init__(self, descriptor=None, *, keys=None, keyed=None):
-        self.setinitial("keys", keys)
-        self.setinitial("keyed", keyed)
-        super().__init__(descriptor)
-
-    @Metadata.property
-    def data_keys(self):
-        """
-        Returns:
-            str[]?: keys
-        """
-        return self.get("keys")
-
-    @Metadata.property
-    def keyed(self):
-        """
-        Returns:
-            bool: keyed
-        """
-        return self.get("keyed", False)
-
-    # Expand
-
-    def expand(self):
-        """Expand metadata"""
-        self.setdefault("keyed", self.keyed)
-
-    # Metadata
-
-    metadata_profile = {  # type: ignore
-        "type": "object",
-        "additionalProperties": False,
-        "properties": {
-            "keys": {"type": "array"},
-            "keyed": {"type": "boolean"},
-        },
-    }
-
-
-# Parser
+from ...exception import FrictionlessException
+from ...parser import Parser
+from ... import errors
 
 
 class InlineParser(Parser):
